requirements.txt
# Core dependencies for Agentic Researcher RAG System
flask>=2.3.0
flask-socketio>=5.3.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-tavily>=0.1.0
langgraph>=0.1.0
openai>=1.0.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pydantic>=2.0.0
tiktoken>=0.5.0
httpx>=0.24.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0